Add helper role for encoding configurations on template.
config-encoders, or config-encoder-filters, is a helper role to encode
a structure (such as a dictionary or list) into a given format: JSON,
YAML, INI, etc.

This is useful for configuration files because you can define the entire
configuration file in your defaults, and any overrides, and it will all
be written out to your template without needing to change the template.
Thus, the structure of the data you encode informs the structure of the
output, instead of the other way around.